memcpy was updated

+++ ft_memcmp.c
@@ -6,13 +6,13 @@
 /*   By: hatalhao <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/03 17:24:52 by hatalhao          #+#    #+#             */
-/*   Updated: 2023/11/03 21:37:23 by hatalhao         ###   ########.fr       */
+/*   Updated: 2023/11/04 20:32:19 by hatalhao         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_memcmp(const void *s1, const void *s2, unsigned int n)
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
 	unsigned char	*str1;
 	unsigned char	*str2;
